% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateno.R
\name{assert_stateno_format_valid}
\alias{assert_stateno_format_valid}
\title{Confirm no stateno invalidly-formatted}
\usage{
assert_stateno_format_valid(df, stateno_var = stateno)
}
\arguments{
\item{df}{A data frame}

\item{stateno_var}{name (unquoted) of a variable in \code{df} containing stateno}
}
\value{
The input data frame \code{df} (if no invalid stateno are found)
}
\description{
Throws an error if the specified variable in the specified
data frame contains any values that do no adhere to Washington State's
formatting conventions for stateno and so appear to be invalid
}
\examples{
read_ehars_person(col_select = stateno) |> assert_stateno_format_valid() |> View() # opens listing of stateno in R's Data Viewer
data.frame(stateno = c('12345')) |> assert_stateno_format_valid() |> View() # generates an error

}
