% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-basic.R
\name{calc_age_groups}
\alias{calc_age_groups}
\title{Calculate age groups at diagnosis and currently}
\usage{
calc_age_groups(df, cur_age_as_of = as.Date(getOption("kcHARS.ehars_dt")))
}
\arguments{
\item{df}{A data.frame containing (at least) eHARS variables
\code{hiv_aids_age_yrs} and dob}

\item{cur_age_as_of}{A Date value specifying the date as of which to calculate
each person's "current" age. Defaults to the date associated with the eHARS
data being used.}
}
\value{
Input data frame \code{df} with the addition of variables
\code{hiv_aids_age_group} and \code{cur_age_group}
}
\description{
Calculate age groups at diagnosis and currently
}
\examples{
read_ehars_person(col_select = c(dob, hiv_aids_age_yrs)) |> calc_age_groups() |> View()
}
