% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-basic.R
\name{calc_race_eth_vars}
\alias{calc_race_eth_vars}
\title{Calculate non-mutually exclusive race/ethnicity variables}
\usage{
calc_race_eth_vars(df)
}
\arguments{
\item{df}{A data.frame containing (at least) eHARS variables race1-race5 and ethnicity1}
}
\value{
Input data frame \code{df} with the addition of variables (1)
natam_multi, (2) asian_multi, (3) black_multi, (4) pacisl_multi, (5)
white_multi, and (6) latino
}
\description{
Calculate non-mutually exclusive race/ethnicity variables
}
\examples{
read_ehars_person(col_select = c(matches("race[0-9]"), all_of(c("ethnicity1")))) |> dplyr::distinct() |> calc_race_eth_vars() |> View()
}
