% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehars_dt.as.Date.R
\name{ehars_dt.as.Date}
\alias{ehars_dt.as.Date}
\title{Convert eHARS-style dates from character to date values}
\usage{
ehars_dt.as.Date(x, ..to = "15", ....to = "0630")
}
\arguments{
\item{x}{A character vector of eHARS-style date values: 'YYYYMMDD' with
missing day/month+day/date specified as '..'/'....'/'........'.}

\item{..to}{Two-digit (zero-padded, as needed) character value specifying the
day to impute, where missing. Defaults to '15'.}

\item{....to}{Four-digit (zero-padded, as needed) character value specifying
the month+day to impute, where missing. Defaults to '0630'.}
}
\value{
A vector of class \code{Date}
}
\description{
Convert eHARS-style dates ('YYYYMMDD' with missing
day/month+day/date specified as '..'/'....'/'........') from character to
date values. '........' are converted to NAs.
}
\examples{
ehars_dt.as.Date("20191231")
ehars_dt.as.Date("201901..")
ehars_dt.as.Date("201901..", ..to = '01')
ehars_dt.as.Date("2019....")
ehars_dt.as.Date("2019....", ....to = '0615')
}
