% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateno.R
\name{fix_5digit_stateno}
\alias{fix_5digit_stateno}
\title{Re-add leading 0 to stateno where dropped (e.g., by Excel)}
\usage{
fix_5digit_stateno(df, stateno_var = stateno)
}
\arguments{
\item{df}{A data frame}

\item{stateno_var}{name (unquoted) of a variable in \code{df} containing stateno}
}
\value{
The input data frame \code{df} with the leading 0 added back to 5-digit
values of column \code{stateno_var}
}
\description{
Re-add leading 0 to stateno where dropped (e.g., by Excel)
}
\examples{
data.frame(stateno = c("12345")) |> fix_5digit_stateno() # leading 0 added
}
