% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{group_race_eth_multi_vars}
\alias{group_race_eth_multi_vars}
\title{Group breakdowns by non-mutually exclusive race/ethnicity variables (e.g.,
natam_multi, black_multi, latino)}
\usage{
group_race_eth_multi_vars(x)
}
\arguments{
\item{x}{gtsummary object of class 'tbl_summary' containing breakdowns by
variables \code{black_multi}, \code{natam_multi}, \code{asian_multi}, \code{black_multi},
\code{latino}, \code{pacisl_multi}, and \code{white_multi}}
}
\value{
a gtsummary table modified such that breakdowns by non-mutually
exclusive race/ethnicity variables display as though they are a breakdown by
a single race/ethnicity variable, with a footnote clarifying that categories
are not mutually exclusive
}
\description{
Group breakdowns by non-mutually exclusive race/ethnicity variables (e.g.,
natam_multi, black_multi, latino)
}
\examples{
x <- kcHARS::read_eoq_append(
date = "2026-07-01",
  col_select = c(ethnicity1, matches("race[0-9]"))
  ) |>
    kcHARS::calc_race_eth_vars() |>
    select(-ethnicity1, -matches("race[0-9]"))

y <- x |> Misc.SHH.f::labelled_to_factor() |> gtsummary::tbl_summary()
y # what the table looks like BEFORE applying group_race_eth_multi_vars()
y |> group_race_eth_multi_vars() # and after
}
