% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehars_dt.as.Date.R
\name{is.ehars_dt}
\alias{is.ehars_dt}
\title{Assert a vector contains eHARS-tyle dates (YYYYMMDD with .. for missing date
parts)}
\usage{
is.ehars_dt(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A single boolean value indicating whether all elements of \code{x} are
eHARS dates
}
\description{
Assert a vector contains eHARS-tyle dates (YYYYMMDD with .. for missing date
parts)
}
\examples{
is.ehars_dt(c("20260101", "202601..", "2026....", "........")) # TRUE
is.ehars_dt(c("20260..1")) # NOTE: ignores misplaced .. (since these do occur in eHARS on rare occasions)
is.ehars_dt(c("2026-01-01", "2026-02-02")) # FALSE
}
