% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehars.R
\name{read_ehars_person}
\alias{read_ehars_person}
\title{Read eHARS' person-based dataset (AKA the person view) into R}
\usage{
read_ehars_person(...)
}
\arguments{
\item{...}{Optional parameters passed to haven::read_sas()}
}
\value{
a data frame
}
\description{
Read eHARS' person-based dataset (AKA the person view) into R
}
\examples{
read_ehars_person(n_max = 10, col_select = c("stateno", "hiv_aids_dx_dt", "trans_categ"))
}
