% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoq.R
\name{read_eoq}
\alias{read_eoq}
\title{Read EOQ datasets (eoqrep_\if{html}{\out{<YYYYMM>}}_newcr) into R}
\usage{
read_eoq(YYYYMM = NULL, ...)
}
\arguments{
\item{YYYYMM}{6-digit character value specifying the year and month of the
dataset (e.g., "202607"). If NULL (the default), the latest EOQ dataset is
read}

\item{...}{Optional parameters passed to haven::read_sas()}
}
\value{
A data frame
}
\description{
Read EOQ datasets (eoqrep_\if{html}{\out{<YYYYMM>}}_newcr) into R
}
\examples{
read_eoq(col_select = c(enter_dt)) |> dplyr::pull(enter_dt) |> max()
read_eoq(YYYYMM = "202607", col_select = c(enter_dt)) |> dplyr::pull(enter_dt) |> max()
read_eoq(YYYYMM = "202511", col_select = c(enter_dt)) |> dplyr::pull(enter_dt) |> max()
read_eoq(YYYYMM = "190001", col_select = c(enter_dt)) |> dplyr::pull(enter_dt) |> max()
}
