% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoq_append.R
\name{read_eoq_append}
\alias{read_eoq_append}
\title{Read EOQ datasets (eoqrep_\if{html}{\out{<YYYYMM>}}_newcr) into R}
\usage{
read_eoq_append(date = NULL, ...)
}
\arguments{
\item{date}{Character string specifying the date of the eoq_append dataset to
read (formatted as YYYY-MM-DD)}

\item{...}{Optional parameters passed to haven::read_sas()}
}
\value{
A data frame
}
\description{
Read EOQ datasets (eoqrep_\if{html}{\out{<YYYYMM>}}_newcr) into R
}
\examples{
read_eoq_append(col_select = c(enter_dt)) |> dplyr::pull(enter_dt) |> max()
read_eoq_append(date = "2026-07-01", col_select = c(enter_dt)) |> dplyr::pull(enter_dt) |> max()
read_eoq_append(date = "2024-04-08", col_select = c(enter_dt)) |> dplyr::pull(enter_dt) |> max()
read_eoq_append(date = "1900-07-01", col_select = c(enter_dt)) |> dplyr::pull(enter_dt) |> max()
}
