% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-basic.R
\name{calc_gender}
\alias{calc_gender}
\title{Calculate 3- or 5-category gender variable}
\usage{
calc_gender(df, n_cat = 5)
}
\arguments{
\item{df}{A data frame containing (at least) eHARS variables sex/birth_sex,
LF_TRANS, LF_GENDER, and LF_GENDER_OTHER}

\item{n_cat}{Number of categories to output; if 3, then all non-cisgender
people included in a single 'Gender Diverse' category}
}
\value{
Input data frame \code{df} with the addition of variables (1) gender, and
(2) transgender_ever
}
\description{
Calculate 3- or 5-category gender variable
}
\examples{
read_ehars_person(col_select = c(sex, LF_TRANS, LF_GENDER, LF_GENDER_OTHER)) |> dplyr::distinct() |> calc_gender() |> View()
}
