% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_KingCo_region}
\alias{calc_KingCo_region}
\title{Calculate King County region}
\usage{
calc_KingCo_region(
  df,
  region_var,
  zip_cd_var,
  county_name_var,
  state_cd_var,
  hml_var = NA_character_
)
}
\arguments{
\item{df}{A data frame}

\item{region_var}{quoted name of the output variable specifying region of King County}

\item{zip_cd_var}{quoted name of the variable specifying zip code of residence; can be >5 digits but only the first five are used}

\item{county_name_var}{quoted name of the variable specifying county of residence (e.g., 'King Co.')}

\item{state_cd_var}{quoted name of the variable specifying the 2-letter code of the state of residence (e.g., 'WA')}

\item{hml_var}{optional 0/1 indicator for whether the person was homeless/unstably housed (H/UH); if specified, H/UH appear in their own category}
}
\value{
Input data frame \code{df} with the addition of region_var
}
\description{
Calculate King County region
}
\examples{
read_ehars_person(col_select = c()) |> calc_KingCo_region(region_var = "res_region_inc", zip_cd_var = "rsd_zip_cd", county_name_var = "rsd_county_name", state_cd_var = "rsd_state_cd") |>  View()
}
