% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-basic.R
\name{calc_nativity}
\alias{calc_nativity}
\title{Calculate nativity (US-born vs foreign-born vs unknown)}
\usage{
calc_nativity(df)
}
\arguments{
\item{df}{A data.frame containing (at least) eHARS variable birth_country_cd}
}
\value{
Input data frame \code{df} with the addition of variable \code{nativity}
}
\description{
Calculate nativity (US-born vs foreign-born vs unknown)
}
\examples{
read_ehars_person(col_select = birth_country_cd) |> dplyr::distinct() |> calc_nativity() |> View()
}
