% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-basic.R
\name{calc_trans_categ}
\alias{calc_trans_categ}
\title{Calculate transmission category variables}
\usage{
calc_trans_categ(df)
}
\arguments{
\item{df}{a data frame containing (at least) eHARS variables \code{trans_categ},
\code{sex}, \code{sex_male}, \code{sex_female}, \code{idu}, and variable \code{gender} calculated by
\code{kcHARS::calc_gender()}}
}
\value{
Input data frame \code{df} with the addition of variables
\code{trans_categ_exp} and \code{msm}
}
\description{
Calculate transmission category variables
}
\examples{
read_ehars_person(col_select = dplyr::any_of(c(
  'trans_categ',
  'sex',
  'sex_male',
  'sex_female',
  'idu',
  "LF_TRANS",
  "LF_GENDER",
  "LF_GENDER_OTHER"
))
) |> distinct() |> calc_gender() |> calc_trans_categ() |> View()
}
