% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehars.R
\name{ehars_root_dir}
\alias{ehars_root_dir}
\title{Return the path to the root directory for eHARS files}
\usage{
ehars_root_dir()
}
\value{
A character value specifying the path to the root directory for
eHARS files
}
\description{
Return the path to the root directory for eHARS files
}
