% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ehars.R
\name{read_ehars_doc}
\alias{read_ehars_doc}
\title{Read an eHARS document-based dataset (e.g., document, lab) into R}
\usage{
read_ehars_doc(tbl, ...)
}
\arguments{
\item{tbl}{character value specifying the name of one of eHARS document-based
tables}

\item{...}{Optional parameters passed to haven::read_sas()}
}
\value{
a data.frame
}
\description{
Read an eHARS document-based dataset (e.g., document, lab) into R
}
\examples{
read_ehars_doc("document", n_max = 10)
read_ehars_doc("document", n_max = 10, col_select = c("ehars_uid", "document_uid", "document_type_cd", "enter_dt"))
}
